WIP: Implement wait_for_port equivalent from conu
Sadly this does not work at the moment as we support rootless podman. The issue
is that for rootless containers we do not get an IP, only the exposed port on
the host. This port is always open, always accepting connections and will also
happily receive data without an error even if nothing is listening inside the
container :-/

See also: https://github.com/dcermak/pytest_container/issues/143