Fast fail containers with a HEALTHCHECK that are not running
Currently a container that has failed to launch completely and is dead will
still be checked for the whole timeout period whether it is healthy. This is
pointless, as it will not be restarted by itself. Hence we fail early now.

This fixes https://github.com/dcermak/pytest_container/issues/207